refactor: break down playermetrics into smaller components
